rearange and consolidate a bit
